chore: initial lh test
