+++ .eslintignore
@@ -1,1 +1,3 @@
 helix-importer-ui
+scripts/gasp-3_11_3-min.js
+scripts/vimeo-api-player.js
