+++ scripts/scripts.js
@@ -28,6 +28,12 @@ function decoratePageTheme() {
   }
 }
 
+function decorateIconPlaceholders(main) {
+  main.querySelectorAll('span.icon').forEach((span) => {
+    span.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg"></svg>';
+  });
+}
+
 /**
  * Decorates the main element.
  * @param {Element} main The main element
@@ -36,9 +42,9 @@ function decoratePageTheme() {
 export function decorateMain(main) {
   // hopefully forward compatible button decoration
   decorateButtons(main);
-  decorateIcons(main);
   decorateSections(main);
   decorateBlocks(main);
+  decorateIconPlaceholders(main);
 }
 
 /**
@@ -71,18 +77,24 @@ export function addFavIcon(href) {
   }
 }
 
-// add external script before body tag ends to ensure availability
-export function addLibScriptBeforeBodyEndTag(cdnLink) {
-  const cdnScript = document.createElement('script');
-  cdnScript.src = cdnLink;
-  document.body.append(cdnScript);
-}
+function addFadeUp() {
+  const observerOptions = {
+    threshold: 0.10,
+    rootMargin: '-10px 0px -10px 0px',
+  };
 
-// batch add external scripts
-function addExternalCDNScriptsLazy() {
-  const cdnLinks = [`${window.hlx.codeBasePath}/scripts/gasp-3_11_3-min.js`];
-  cdnLinks.forEach((link) => {
-    addLibScriptBeforeBodyEndTag(link);
+  const observer = new IntersectionObserver((entries) => {
+    entries.forEach((entry) => {
+      if (entry.isIntersecting) {
+        entry.target.classList.add('in-view');
+        observer.unobserve(entry.target);
+      }
+    });
+  }, observerOptions);
+
+  const sections = Array.from(document.getElementsByClassName('fadeup'));
+  sections.forEach((section) => {
+    observer.observe(section);
   });
 }
 
@@ -90,6 +102,7 @@ function addExternalCDNScriptsLazy() {
  * loads everything that doesn't need to be delayed.
  */
 async function loadLazy(doc) {
+  addFadeUp();
   const main = doc.querySelector('main');
   await loadBlocks(main);
 
@@ -97,9 +110,9 @@ async function loadLazy(doc) {
   const element = hash ? main.querySelector(hash) : false;
   if (hash && element) element.scrollIntoView();
 
-  addExternalCDNScriptsLazy();
   await loadHeader(doc.querySelector('header'));
   await loadFooter(doc.querySelector('footer'));
+  await decorateIcons(main);
 
   decoratePageTheme();
 
@@ -120,20 +133,6 @@ function loadDelayed() {
   // load anything that can be postponed to the latest here
 }
 
-const observerOptions = {
-  threshold: 0.10,
-  rootMargin: '-10px 0px -10px 0px',
-};
-
-const observer = new IntersectionObserver((entries) => {
-  entries.forEach((entry) => {
-    if (entry.isIntersecting) {
-      entry.target.classList.add('in-view');
-      observer.unobserve(entry.target);
-    }
-  });
-}, observerOptions);
-
 // Window Resize Handler
 let resizeTimer;
 const resizeCompleteEvent = new CustomEvent('resizeComplete', (e) => {
@@ -149,16 +148,10 @@ window.addEventListener('resize', () => {
 
 async function loadPage() {
   // need vimeo available for vimeo block
-  addLibScriptBeforeBodyEndTag('/scripts/vimeo-api-player.js');
   await loadEager(document);
   await loadLazy(document);
   addNextSectionArrowButton();
   loadDelayed();
 }
 
-loadPage().then(() => {
-  const sections = Array.from(document.getElementsByClassName('fadeup'));
-  sections.forEach((section) => {
-    observer.observe(section);
-  });
-});
+loadPage();
